Move `/locking` and `/campaign` routes under `/community`
