+++ src/routes/locking/locking.controller.spec.ts
@@ -4,46 +4,20 @@ import { INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { TestAppProvider } from '@/__tests__/test-app.provider';
 import { AppModule } from '@/app.module';
-import { IConfigurationService } from '@/config/configuration.service.interface';
 import { TestCacheModule } from '@/datasources/cache/__tests__/test.cache.module';
 import { CacheModule } from '@/datasources/cache/cache.module';
 import { TestNetworkModule } from '@/datasources/network/__tests__/test.network.module';
 import { NetworkModule } from '@/datasources/network/network.module';
-import {
-  INetworkService,
-  NetworkService,
-} from '@/datasources/network/network.service.interface';
 import { TestLoggingModule } from '@/logging/__tests__/test.logging.module';
 import { RequestScopedLoggingModule } from '@/logging/logging.module';
 import configuration from '@/config/entities/__tests__/configuration';
-import { NetworkResponseError } from '@/datasources/network/entities/network.error.entity';
-import { pageBuilder } from '@/domain/entities/__tests__/page.builder';
-import {
-  lockEventItemBuilder,
-  unlockEventItemBuilder,
-  withdrawEventItemBuilder,
-  toJson as lockingEventToJson,
-} from '@/domain/locking/entities/__tests__/locking-event.builder';
-import { LockingEvent } from '@/domain/locking/entities/locking-event.entity';
 import { TestAccountDataSourceModule } from '@/datasources/account/__tests__/test.account.datasource.module';
 import { AccountDataSourceModule } from '@/datasources/account/account.datasource.module';
-import { getAddress } from 'viem';
-import { rankBuilder } from '@/domain/locking/entities/__tests__/rank.builder';
-import { PaginationData } from '@/routes/common/pagination/pagination.data';
 import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queues-api.module';
 import { QueuesApiModule } from '@/datasources/queues/queues-api.module';
-import {
-  campaignBuilder,
-  toJson as campaignToJson,
-} from '@/domain/locking/entities/__tests__/campaign.builder';
-import { Campaign } from '@/domain/locking/entities/campaign.entity';
-import { CampaignRank } from '@/domain/locking/entities/campaign-rank.entity';
-import { campaignRankBuilder } from '@/domain/locking/entities/__tests__/campaign-rank.builder';
 
 describe('Locking (Unit)', () => {
   let app: INestApplication;
-  let lockingBaseUri: string;
-  let networkService: jest.MockedObjectDeep<INetworkService>;
 
   beforeEach(async () => {
     jest.resetAllMocks();
@@ -63,10 +37,6 @@ describe('Locking (Unit)', () => {
       .useModule(TestQueuesApiModule)
       .compile();
 
-    const configurationService = moduleFixture.get(IConfigurationService);
-    lockingBaseUri = configurationService.get('locking.baseUri');
-    networkService = moduleFixture.get(NetworkService);
-
     app = await new TestAppProvider().provide(moduleFixture);
     await app.init();
   });
@@ -75,668 +45,70 @@ describe('Locking (Unit)', () => {
     await app.close();
   });
 
-  describe('GET campaign', () => {
-    it('should get the campaign', async () => {
-      const campaign = campaignBuilder().build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns/${campaign.campaignId}`:
-            return Promise.resolve({ data: campaign, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/campaigns/${campaign.campaignId}`)
-        .expect(200)
-        .expect(campaignToJson(campaign) as Campaign);
-    });
-
-    it('should get the list of campaigns', async () => {
-      const campaignsPage = pageBuilder<Campaign>()
-        .with('results', [campaignBuilder().build()])
-        .with('count', 1)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns`:
-            return Promise.resolve({ data: campaignsPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/campaigns`)
-        .expect(200)
-        .expect({
-          count: 1,
-          next: null,
-          previous: null,
-          results: campaignsPage.results.map(campaignToJson),
-        });
-    });
-
-    it('should validate the list of campaigns', async () => {
-      const invalidCampaigns = [{ invalid: 'campaign' }];
-      const campaignsPage = pageBuilder()
-        .with('results', invalidCampaigns)
-        .with('count', 1)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns`:
-            return Promise.resolve({ data: campaignsPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/campaigns`)
-        .expect(500)
-        .expect({
-          statusCode: 500,
-          message: 'Internal server error',
-        });
-    });
-
-    it('should forward the pagination parameters', async () => {
-      const limit = faker.number.int({ min: 1, max: 10 });
-      const offset = faker.number.int({ min: 1, max: 10 });
-      const campaignsPage = pageBuilder<Campaign>()
-        .with('results', [campaignBuilder().build()])
-        .with('count', 1)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns`:
-            return Promise.resolve({ data: campaignsPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(
-          `/v1/locking/campaigns?cursor=limit%3D${limit}%26offset%3D${offset}`,
-        )
-        .expect(200)
-        .expect({
-          count: 1,
-          next: null,
-          previous: null,
-          results: campaignsPage.results.map(campaignToJson),
-        });
-
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${lockingBaseUri}/api/v1/campaigns`,
-        networkRequest: {
-          params: {
-            limit,
-            offset,
-          },
-        },
-      });
-    });
-
-    it('should forward errors from the service', async () => {
-      const statusCode = faker.internet.httpStatusCode({
-        types: ['clientError', 'serverError'],
-      });
-      const errorMessage = faker.word.words();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns`:
-            return Promise.reject(
-              new NetworkResponseError(
-                new URL(`${lockingBaseUri}/api/v1/campaigns`),
-                {
-                  status: statusCode,
-                } as Response,
-                { message: errorMessage, status: statusCode },
-              ),
-            );
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/campaigns`)
-        .expect(statusCode)
-        .expect({
-          message: errorMessage,
-          code: statusCode,
-        });
-    });
-  });
-
-  describe('GET leaderboard by campaign', () => {
-    it('should get the leaderboard by campaign', async () => {
-      const campaign = campaignBuilder().build();
-      const campaignRankPage = pageBuilder<CampaignRank>()
-        .with('results', [
-          campaignRankBuilder().build(),
-          campaignRankBuilder().build(),
-        ])
-        .with('count', 2)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns/${campaign.campaignId}/leaderboard`:
-            return Promise.resolve({ data: campaignRankPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/campaigns/${campaign.campaignId}/leaderboard`)
-        .expect(200)
-        .expect({
-          count: 2,
-          next: null,
-          previous: null,
-          results: campaignRankPage.results,
-        });
-    });
-
-    it('should validate the response', async () => {
-      const campaign = campaignBuilder().build();
-      const invalidCampaignRanks = [{ invalid: 'campaignRank' }];
-      const campaignRankPage = pageBuilder()
-        .with('results', invalidCampaignRanks)
-        .with('count', invalidCampaignRanks.length)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns/${campaign.campaignId}/leaderboard`:
-            return Promise.resolve({ data: campaignRankPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/campaigns/${campaign.campaignId}/leaderboard`)
-        .expect(500)
-        .expect({
-          statusCode: 500,
-          message: 'Internal server error',
-        });
-    });
-
-    it('should forward the pagination parameters', async () => {
-      const limit = faker.number.int({ min: 1, max: 10 });
-      const offset = faker.number.int({ min: 1, max: 10 });
-      const campaign = campaignBuilder().build();
-      const campaignRankPage = pageBuilder<CampaignRank>()
-        .with('results', [
-          campaignRankBuilder().build(),
-          campaignRankBuilder().build(),
-        ])
-        .with('count', 2)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns/${campaign.campaignId}/leaderboard`:
-            return Promise.resolve({ data: campaignRankPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(
-          `/v1/locking/campaigns/${campaign.campaignId}/leaderboard?cursor=limit%3D${limit}%26offset%3D${offset}`,
-        )
-        .expect(200)
-        .expect({
-          count: 2,
-          next: null,
-          previous: null,
-          results: campaignRankPage.results,
-        });
-
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${lockingBaseUri}/api/v1/campaigns/${campaign.campaignId}/leaderboard`,
-        networkRequest: {
-          params: {
-            limit,
-            offset,
-          },
-        },
-      });
-    });
-
-    it('should forward errors from the service', async () => {
-      const campaign = campaignBuilder().build();
-      const statusCode = faker.internet.httpStatusCode({
-        types: ['clientError', 'serverError'],
-      });
-      const errorMessage = faker.word.words();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/campaigns/${campaign.campaignId}/leaderboard`:
-            return Promise.reject(
-              new NetworkResponseError(
-                new URL(url),
-                {
-                  status: statusCode,
-                } as Response,
-                { message: errorMessage, status: statusCode },
-              ),
-            );
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/campaigns/${campaign.campaignId}/leaderboard`)
-        .expect(statusCode)
-        .expect({
-          message: errorMessage,
-          code: statusCode,
-        });
-    });
-  });
-
-  describe('GET rank', () => {
-    it('should get the rank', async () => {
-      const rank = rankBuilder().build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/leaderboard/${rank.holder}`:
-            return Promise.resolve({ data: rank, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/leaderboard/rank/${rank.holder}`)
-        .expect(200)
-        .expect(rank);
-    });
-
-    it('should validate the Safe address in URL', async () => {
-      const safeAddress = faker.string.alphanumeric();
+  describe('GET /locking/leaderboard/rank/:safeAddress', () => {
+    it('should return 302 and redirect to the new endpoint', async () => {
+      const safeAddress = faker.finance.ethereumAddress();
 
       await request(app.getHttpServer())
         .get(`/v1/locking/leaderboard/rank/${safeAddress}`)
-        .expect(422)
-        .expect({
-          statusCode: 422,
-          code: 'custom',
-          message: 'Invalid address',
-          path: [],
-        });
-    });
-
-    it('should validate the response', async () => {
-      const safeAddress = getAddress(faker.finance.ethereumAddress());
-      const rank = { invalid: 'rank' };
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/leaderboard/${safeAddress}`:
-            return Promise.resolve({ data: rank, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/leaderboard/rank/${safeAddress}`)
-        .expect(500)
-        .expect({
-          statusCode: 500,
-          message: 'Internal server error',
-        });
-    });
-
-    it('should forward an error from the service', async () => {
-      const safeAddress = getAddress(faker.finance.ethereumAddress());
-      const statusCode = faker.internet.httpStatusCode({
-        types: ['clientError', 'serverError'],
-      });
-      const errorMessage = faker.word.words();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/leaderboard/${safeAddress}`:
-            return Promise.reject(
-              new NetworkResponseError(
-                new URL(`${lockingBaseUri}/api/v1/leaderboard/${safeAddress}`),
-                {
-                  status: statusCode,
-                } as Response,
-                { message: errorMessage, status: statusCode },
-              ),
-            );
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/leaderboard/rank/${safeAddress}`)
-        .expect(statusCode)
-        .expect({
-          message: errorMessage,
-          code: statusCode,
+        .expect(308)
+        .expect((res) => {
+          expect(res.get('location')).toBe(
+            `/v1/community/locking/${safeAddress}/rank`,
+          );
         });
     });
   });
 
-  describe('GET leaderboard', () => {
-    it('should get the leaderboard', async () => {
-      const leaderboard = pageBuilder()
-        .with('results', [rankBuilder().build()])
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/leaderboard`:
-            return Promise.resolve({ data: leaderboard, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
+  describe('GET /locking/leaderboard', () => {
+    it('should return 302 and redirect to the new endpoint', async () => {
       await request(app.getHttpServer())
-        .get(`/v1/locking/leaderboard`)
-        .expect(200)
-        .expect(({ body }) => {
-          expect(body).toEqual({
-            count: leaderboard.count,
-            next: expect.any(String),
-            previous: expect.any(String),
-            results: leaderboard.results,
-          });
-        });
-
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${lockingBaseUri}/api/v1/leaderboard`,
-        networkRequest: {
-          params: {
-            limit: PaginationData.DEFAULT_LIMIT,
-            offset: PaginationData.DEFAULT_OFFSET,
-          },
-        },
-      });
-    });
-
-    it('should forward the pagination parameters', async () => {
-      const limit = faker.number.int({ min: 1, max: 10 });
-      const offset = faker.number.int({ min: 1, max: 10 });
-      const leaderboard = pageBuilder()
-        .with('results', [rankBuilder().build()])
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/leaderboard`:
-            return Promise.resolve({ data: leaderboard, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(
-          `/v1/locking/leaderboard?cursor=limit%3D${limit}%26offset%3D${offset}`,
-        )
-        .expect(200)
-        .expect(({ body }) => {
-          expect(body).toEqual({
-            count: leaderboard.count,
-            next: expect.any(String),
-            previous: expect.any(String),
-            results: leaderboard.results,
-          });
-        });
-
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${lockingBaseUri}/api/v1/leaderboard`,
-        networkRequest: {
-          params: {
-            limit,
-            offset,
-          },
-        },
-      });
-    });
-
-    it('should validate the response', async () => {
-      const leaderboard = pageBuilder()
-        .with('results', [{ invalid: 'rank' }])
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/leaderboard`:
-            return Promise.resolve({ data: leaderboard, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/leaderboard`)
-        .expect(500)
-        .expect({
-          statusCode: 500,
-          message: 'Internal server error',
+        .get('/v1/locking/leaderboard')
+        .expect(308)
+        .expect((res) => {
+          expect(res.get('location')).toBe('/v1/community/locking/leaderboard');
         });
     });
 
-    it('should forward an error from the service', async () => {
-      const statusCode = faker.internet.httpStatusCode({
-        types: ['clientError', 'serverError'],
-      });
-      const errorMessage = faker.word.words();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/leaderboard`:
-            return Promise.reject(
-              new NetworkResponseError(
-                new URL(`${lockingBaseUri}/api/v1/leaderboard`),
-                {
-                  status: statusCode,
-                } as Response,
-                { message: errorMessage, status: statusCode },
-              ),
-            );
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
+    it('should return 302 and redirect to the new endpoint with cursor', async () => {
+      const cursor = 'limit%3Daa%26offset%3D2';
 
       await request(app.getHttpServer())
-        .get(`/v1/locking/leaderboard`)
-        .expect(statusCode)
-        .expect({
-          message: errorMessage,
-          code: statusCode,
+        .get(`/v1/locking/leaderboard/?cursor=${cursor}`)
+        .expect(308)
+        .expect((res) => {
+          expect(res.get('location')).toBe(
+            `/v1/community/locking/leaderboard/?cursor=${cursor}`,
+          );
         });
     });
   });
 
-  describe('GET locking history', () => {
-    it('should get locking history', async () => {
+  describe('GET /locking/:safeAddress/history', () => {
+    it('should return 302 and redirect to the new endpoint', async () => {
       const safeAddress = faker.finance.ethereumAddress();
-      const lockingHistory = [
-        lockEventItemBuilder().build(),
-        unlockEventItemBuilder().build(),
-        withdrawEventItemBuilder().build(),
-      ];
-      const lockingHistoryPage = pageBuilder<LockingEvent>()
-        .with('results', lockingHistory)
-        .with('count', lockingHistory.length)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          // Service will have checksummed address
-          case `${lockingBaseUri}/api/v1/all-events/${getAddress(safeAddress)}`:
-            return Promise.resolve({ data: lockingHistoryPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
 
       await request(app.getHttpServer())
         .get(`/v1/locking/${safeAddress}/history`)
-        .expect(200)
-        .expect({
-          count: lockingHistoryPage.count,
-          next: null,
-          previous: null,
-          results: lockingHistoryPage.results.map(lockingEventToJson),
+        .expect(308)
+        .expect((res) => {
+          expect(res.get('location')).toBe(
+            `/v1/community/locking/${safeAddress}/history`,
+          );
         });
-
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${lockingBaseUri}/api/v1/all-events/${getAddress(safeAddress)}`,
-        networkRequest: {
-          params: {
-            limit: PaginationData.DEFAULT_LIMIT,
-            offset: PaginationData.DEFAULT_OFFSET,
-          },
-        },
-      });
     });
 
-    it('should forward the pagination parameters', async () => {
+    it('should return 302 and redirect to the new endpoint with cursor', async () => {
       const safeAddress = faker.finance.ethereumAddress();
-      const limit = faker.number.int({ min: 1, max: 10 });
-      const offset = faker.number.int({ min: 1, max: 10 });
-      const lockingHistory = [
-        lockEventItemBuilder().build(),
-        unlockEventItemBuilder().build(),
-        withdrawEventItemBuilder().build(),
-      ];
-      const lockingHistoryPage = pageBuilder<LockingEvent>()
-        .with('results', lockingHistory)
-        .with('count', lockingHistory.length)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          // Service will have checksummed address
-          case `${lockingBaseUri}/api/v1/all-events/${getAddress(safeAddress)}`:
-            return Promise.resolve({ data: lockingHistoryPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
+      const cursor = 'limit%3Daa%26offset%3D2';
 
       await request(app.getHttpServer())
-        .get(
-          `/v1/locking/${safeAddress}/history?cursor=limit%3D${limit}%26offset%3D${offset}`,
-        )
-        .expect(200)
-        .expect({
-          count: lockingHistoryPage.count,
-          next: null,
-          previous: null,
-          results: lockingHistoryPage.results.map(lockingEventToJson),
-        });
-
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${lockingBaseUri}/api/v1/all-events/${getAddress(safeAddress)}`,
-        networkRequest: {
-          params: {
-            limit,
-            offset,
-          },
-        },
-      });
-    });
-
-    it('should validate the Safe address in URL', async () => {
-      const safeAddress = faker.string.alphanumeric();
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/${safeAddress}/history`)
-        .expect(422)
-        .expect({
-          statusCode: 422,
-          code: 'custom',
-          message: 'Invalid address',
-          path: [],
-        });
-    });
-
-    it('should validate the response', async () => {
-      const safeAddress = faker.finance.ethereumAddress();
-      const invalidLockingHistory = [{ invalid: 'value' }];
-      const lockingHistoryPage = pageBuilder()
-        .with('results', invalidLockingHistory)
-        .with('count', invalidLockingHistory.length)
-        .with('previous', null)
-        .with('next', null)
-        .build();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          // Service will have checksummed address
-          case `${lockingBaseUri}/api/v1/all-events/${getAddress(safeAddress)}`:
-            return Promise.resolve({ data: lockingHistoryPage, status: 200 });
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/${safeAddress}/history`)
-        .expect(500)
-        .expect({
-          statusCode: 500,
-          message: 'Internal server error',
-        });
-    });
-
-    it('should forward an error from the service', async () => {
-      const safeAddress = faker.finance.ethereumAddress();
-      const statusCode = faker.internet.httpStatusCode({
-        types: ['clientError', 'serverError'],
-      });
-      const errorMessage = faker.word.words();
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${lockingBaseUri}/api/v1/all-events/${getAddress(safeAddress)}`:
-            return Promise.reject(
-              new NetworkResponseError(
-                new URL(`${lockingBaseUri}/v1/locking/${safeAddress}/history`),
-                {
-                  status: statusCode,
-                } as Response,
-                { message: errorMessage, status: statusCode },
-              ),
-            );
-          default:
-            return Promise.reject(`No matching rule for url: ${url}`);
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/locking/${safeAddress}/history`)
-        .expect(statusCode)
-        .expect({
-          message: errorMessage,
-          code: statusCode,
+        .get(`/v1/locking/${safeAddress}/history/?cursor=${cursor}`)
+        .expect(308)
+        .expect((res) => {
+          expect(res.get('location')).toBe(
+            `/v1/community/locking/${safeAddress}/history/?cursor=${cursor}`,
+          );
         });
     });
   });
